requirements.txt
fastapi
uvicorn
firebase-admin
dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib